separate files into headers and implementation files with pragma guard
