+++ src/page.h
@@ -9,7 +9,7 @@
 //
 //
 //****************************/
-
+#pragma once
 
 #include <stdio.h>
 #include <stdlib.h>
@@ -30,19 +30,7 @@ struct Page {
 //   columns: # of columns " " " " " 
 //
 //***********************************************/
-void pageInit(struct Page* page, int rows, int columns) {
-
-  //Initiate arrays with the size large enough to fill the screen
-  page->lines = malloc( rows * sizeof(char *));
-  page->sizes = malloc( rows * sizeof(int));
- 
-  for(int i = 0; i < rows; i++) { 
-    page->sizes[i] = 0;
-    page->lines[i] = malloc( columns * sizeof(char));
-    page->lines[i][0] = '\0';
-  }
-
-}
+void pageInit(struct Page* page, int rows, int columns);
 
 //***********************************************/
 // insert:
@@ -55,17 +43,7 @@ void pageInit(struct Page* page, int rows, int columns) {
 //   c: character to insert 
 //
 //***********************************************/
-void insert(struct Page* page, int row, int col, char c) {
-  int size = page->sizes[row];
-
-  for(int i = size; i >= col; i--) {
-    page->lines[row][i+1] = page->lines[row][i];
-    if(i == col)
-      page->lines[row][i] = c;
-  }
-
-  page->sizes[row] = size + 1;
-}
+void insert(struct Page* page, int row, int col, char c);
 
 //***********************************************/
 // backspace:
@@ -77,13 +55,4 @@ void insert(struct Page* page, int row, int col, char c) {
 //   col: column # of the position of the character
 //
 //***********************************************/
-void backspace(struct Page* page, int row, int col) {
-  int size = page->sizes[row];
-  
-  for(int i = col; i < size+1; i++) {
-    page->lines[row][i] = page->lines[row][i+1];
-  } 
-
-  page->sizes[row] = size - 1;
-}
-
+void backspace(struct Page* page, int row, int col);
